Change GoogleAssistant to a module
This means adding GoogleAssistant::Assistant to take over the functionality of the old class GoogleAssistant. To keep the API simple, I made it so you can still call:

GoogleAssistant.respond_to(params, response) { |assistant| ... }
